Updates the play and pause buttons with a single play pause button.
Also adds the bootstrap and fontawesome libraries (via includes in index.html)

Finally, there are changes that I'm not committing yet in renderer.js.
I needed to make some overrides to the podcast feed and plug in some default data in order to
test while on an airplane. Will commit this file when those changes are reverted.